+++ examples/stronghold.rs
@@ -3,8 +3,6 @@
 
 //! cargo run --example stronghold --features=stronghold --release
 
-use std::path::PathBuf;
-
 use iota_client::{
     api::GetAddressesBuilder,
     constants::{SHIMMER_COIN_TYPE, SHIMMER_TESTNET_BECH32_HRP},
@@ -17,7 +15,7 @@ use iota_client::{
 async fn main() -> Result<()> {
     let mut stronghold_secret_manager = StrongholdSecretManager::builder()
         .password("some_hopefully_secure_password")
-        .try_build(PathBuf::from("test.stronghold"))?;
+        .try_build("test.stronghold")?;
 
     // This example uses dotenv, which is not safe for use in production
     dotenv::dotenv().ok();